Edit TrafficCard, api

+++ backend/app/models/subscription.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 from uuid import UUID
 
-from sqlalchemy import DateTime, ForeignKey, Integer, String
+from sqlalchemy import BigInteger, DateTime, ForeignKey, String
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.db.models import BaseModel
@@ -44,12 +44,12 @@ class Subscription(BaseModel):
     )
 
     traffic_limit: Mapped[int] = mapped_column(
-        Integer,
+        BigInteger,
         default=1073741824,  # 1 GB
     )
 
     traffic_used: Mapped[int] = mapped_column(
-        Integer,
+        BigInteger,
         default=0,
     )
